+++ RandomSkunk.Results/Operations/AsNonNullable.cs
@@ -31,7 +31,7 @@ public static partial class ResultExtensions
     /// <returns>The equivalant result.</returns>
     public static async Task<Result<T>> AsNonNullable<T>(this Task<Result<T?>> result)
         where T : struct =>
-        (await result).AsNonNullable();
+        (await result.ConfigureAwait(false)).AsNonNullable();
 
     /// <summary>
     /// Gets an equivalent result with a non-nullable type.
@@ -41,5 +41,5 @@ public static partial class ResultExtensions
     /// <returns>The equivalant result.</returns>
     public static async Task<Maybe<T>> AsNonNullable<T>(this Task<Maybe<T?>> result)
         where T : struct =>
-        (await result).AsNonNullable();
+        (await result.ConfigureAwait(false)).AsNonNullable();
 }